moved example dataset into it's own respective file and test file
